Refactor annotation processor to add same defaults for OpenShift as we do for Kubernetes
