+++ common/src/main/java/org/wildfly/test/cloud/common/WildFlyOpenshiftIntegrationTest.java
@@ -19,15 +19,15 @@
 
 package org.wildfly.test.cloud.common;
 
-import org.junit.jupiter.api.extension.ExtendWith;
-
-import java.lang.annotation.Retention;
-import java.lang.annotation.Target;
-
 import static java.lang.annotation.ElementType.ANNOTATION_TYPE;
 import static java.lang.annotation.ElementType.METHOD;
 import static java.lang.annotation.ElementType.TYPE;
 import static java.lang.annotation.RetentionPolicy.RUNTIME;
+
+import java.lang.annotation.Retention;
+import java.lang.annotation.Target;
+
+import org.junit.jupiter.api.extension.ExtendWith;
 
 /**
  * Extends io.dekorate.testing.openshift.annotation.OpenshiftIntegrationTest
