+++ common/src/main/java/org/wildfly/test/cloud/common/WildFlyOpenshiftIntegrationTestConfig.java
@@ -19,10 +19,10 @@
 
 package org.wildfly.test.cloud.common;
 
-import io.dekorate.testing.openshift.config.EditableOpenshiftIntegrationTestConfig;
-
 import java.util.List;
 import java.util.Map;
+
+import io.dekorate.testing.openshift.config.EditableOpenshiftIntegrationTestConfig;
 
 /**
  * @author <a href="mailto:kabir.khan@jboss.com">Kabir Khan</a>
