+++ common/src/main/resources/META-INF/services/javax.annotation.processing.Processor
@@ -1,1 +1,2 @@
 org.wildfly.test.cloud.common.annotation.processor.WildFlyDefaultsKubernetesAnnotationProcessor
+org.wildfly.test.cloud.common.annotation.processor.WildFlyDefaultsOpenshiftAnnotationProcessor
