+++ pom.xml
@@ -65,7 +65,7 @@
         <!-- Kubernetes registry parameters -->
         <wildfly.cloud.test.docker.host>localhost</wildfly.cloud.test.docker.host>
         <wildfly.cloud.test.docker.port>5000</wildfly.cloud.test.docker.port>
-        <wildfly.cloud.test.docker.registry>${wildfly.cloud.test.docker.host}:${wildfly.cloud.test.docker.port}</wildfly.cloud.test.docker.registry>
+        <dekorate.docker.registry>${wildfly.cloud.test.docker.host}:${wildfly.cloud.test.docker.port}</dekorate.docker.registry>
     </properties>
 
     <dependencyManagement>
