+++ tests/datasources/postgresql/src/main/java/org/wildfly/test/cloud/microprofile/datasources/postgresql/PostgresDatasourceEndpoint.java
@@ -19,6 +19,10 @@
 
 package org.wildfly.test.cloud.microprofile.datasources.postgresql;
 
+import static javax.ws.rs.core.MediaType.APPLICATION_JSON;
+
+import java.util.List;
+
 import javax.inject.Inject;
 import javax.ws.rs.Consumes;
 import javax.ws.rs.GET;
@@ -28,9 +32,6 @@ import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
-import java.util.List;
-
-import static javax.ws.rs.core.MediaType.APPLICATION_JSON;
 
 /**
  * @author <a href="mailto:kabir.khan@jboss.com">Kabir Khan</a>
