+++ tests/microprofile/reactive-messaging/strimzi/src/main/java/org/wildfly/test/cloud/microprofile/reactive/messaging/strimzi/ReactiveMessagingWithStrimziApp.java
@@ -19,13 +19,13 @@
 
 package org.wildfly.test.cloud.microprofile.reactive.messaging.strimzi;
 
-import io.dekorate.kubernetes.annotation.Env;
-import io.dekorate.kubernetes.annotation.KubernetesApplication;
+import static io.dekorate.kubernetes.annotation.ImagePullPolicy.Always;
 
 import javax.ws.rs.ApplicationPath;
 import javax.ws.rs.core.Application;
 
-import static io.dekorate.kubernetes.annotation.ImagePullPolicy.Always;
+import io.dekorate.kubernetes.annotation.Env;
+import io.dekorate.kubernetes.annotation.KubernetesApplication;
 
 /**
  * @author <a href="mailto:kabir.khan@jboss.com">Kabir Khan</a>
